@@ -1,8 +1,7 @@
-/** @type {import('postcss-load-config').Config} */
-const config = {
+// eslint-disable-next-line import/no-anonymous-default-export
+export default {
   plugins: {
     tailwindcss: {},
+    autoprefixer: {},
   },
 };
-
-export default config;
